requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Login==0.5.0
Pillow==9.0.0
pytesseract==0.3.8
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.1
